refactor(#6614): Align with new Spring Social versions
