+++ app/server/credential/src/main/java/io/syndesis/server/credential/OAuth2Applicator.java
@@ -19,7 +19,6 @@ import java.util.Date;
 
 import io.syndesis.common.model.connection.Connection;
 
-import org.springframework.boot.autoconfigure.social.SocialProperties;
 import org.springframework.social.connect.ConnectionFactory;
 import org.springframework.social.oauth2.AccessGrant;
