+++ app/server/credential/src/test/java/io/syndesis/server/credential/OAuth2ApplicatorTest.java
@@ -18,7 +18,6 @@ package io.syndesis.server.credential;
 import io.syndesis.common.model.connection.Connection;
 
 import org.junit.Test;
-import org.springframework.boot.autoconfigure.social.SocialProperties;
 import org.springframework.social.oauth2.AccessGrant;
 
 import static org.assertj.core.api.Assertions.assertThat;
